requirements.txt
transformers==4.30.1
torch
torchvision
ftfy
regex
tqdm
beartype
ema_pytorch
vector_quantize_pytorch==1.1.2
monai
xlsxwriter
nibabel
nilearn
numpy
pandas
matplotlib
Pillow
ipywidgets
ipython
h5py
seaborn
accelerate
sentencepiece
peft==0.3.0
rouge-score 
nltk 
bert-score